// enums/filters.ts
export enum PriceFilters {
    UNDER_25 = "Under 25.00",
    FROM_25_TO_50 = "25.00 - 50.00",
    OVER_50 = "Over 50.00",
    UNDER_1000 = "Under 1000.00",
    FROM_1000_TO_1200 = "1000.00 - 1200.00",
    OVER_1200 = "Over 1200.00",
    UNDER_100 = "Under 100.00",
    OVER_100 = "Over 100.00",
    FROM_0_TO_500 = "0.00 - 500.00",
    FROM_500_TO_700 = "500.00 - 700.00",
    FROM_700_TO_3000 = "700.00 - 3000.00",
    UNDER_500 = "Under 500.00",
    OVER_500 = "Over 500.00"
}

// export enum DesktopsPriceFilters {
//     UNDER_1000 = "Under 1000.00",
//     FROM_1000_TO_1200 = "1000.00 - 1200.00",
//     OVER_1200 = "Over 1200.00"
// }

// export enum AccessoriesPriceFilter {
//     UNDER_100 = "Under 100.00",
//     OVER_100 = "Over 100.00"
// }

// export enum CameraPhotoPriceFilter {
//     UNDER_500 = "Under 500.00",
//     OVER_500 = "Over 500.00"
// }

// export enum JewerlyPriceFilter {
//     FROM_0_TO_500 = "0.00 - 500.00",
//     FROM_500_TO_700 = "500.00 - 700.00",
//     FROM_700_TO_3000 = "700.00 - 3000.00"
// }